requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
pgvector==0.3.4
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
aiofiles==24.1.0
numpy==1.26.4
Pillow==10.4.0
huggingface-hub==0.25.2
transformers==4.45.1
torch==2.4.1
timm==1.0.9
sentence-transformers==3.1.1
insightface==0.7.3
onnxruntime==1.19.2
